requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.1.5
django-restframework==0.0.1
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pyscopg2==66.0.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
django-cors-headers